requirement.txt
ipykernel
pandas
matplotlib
wordcloud
scikit-learn
gensim
mlflow
zenml
google-generativeai
xgboost
langdetect